chore: add universal fallbacks for unsupported platform
