--- js/PickerAndroid.macos.js
@@ -1,12 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- * @format
- */
-
-import {UnimplementedView} from 'react-native';
-
-export default UnimplementedView;
