--- js/PickerIOS.android.js
@@ -1,26 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- *
- * This is a controlled component version of RCTPickerIOS
- *
- * @format
- * @flow
- */
-
-'use strict';
-
-import React from 'react';
-import {UnimplementedView} from 'react-native';
-
-class PickerIOS extends React.Component {
-  static Item = UnimplementedView;
-  render() {
-    return <UnimplementedView />;
-  }
-}
-
-export default PickerIOS;
