--- js/PickerIOS.macos.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import {UnimplementedView} from 'react-native';
-
-function PickerIOS() {
-  return <UnimplementedView />;
-}
-
-PickerIOS.Item = UnimplementedView;
-
-export default PickerIOS;
