--- js/PickerIOS.web.js
@@ -1,10 +1,0 @@
-import * as React from 'react';
-import UnimplementedView from 'react-native-web/dist/modules/UnimplementedView';
-
-function PickerIOS() {
-  return <UnimplementedView />;
-}
-
-PickerIOS.Item = UnimplementedView;
-
-export default PickerIOS;
