--- js/PickerMacOS.windows.js
@@ -1,19 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- * @format
- */
-import React from 'react';
-import {UnimplementedView} from 'react-native';
-
-class PickerMacOS extends React.Component {
-  static Item = UnimplementedView;
-  render() {
-    return <UnimplementedView />;
-  }
-}
-
-export default PickerMacOS;
